requirements.txt
pandas~=2.2.1
numpy~=1.26.4
matplotlib~=3.8.4
seaborn~=0.13.2
tqdm~=4.66.2
scipy~=1.13.0
scikit-bio~=0.6.0
pillow~=10.3.0